Split AutoFlow parser into multiple files
